Fixed some typos in MiscInfo.txt and HowToUse.txt Done some Clang-Tidy - Now using range-based for loop - Using better overload of find_last_of - Using const if possible - Using auto if possible - Using raw string literal Removed unused code and variables More consistent paths

+++ src/Gen 4/RomHandler.cpp
@@ -7,16 +7,16 @@ using namespace std;
 using std::filesystem::directory_iterator;
 using std::filesystem::recursive_directory_iterator;
 
-inline string csvGotoNext(string str) {
+inline string csvGotoNext(const string& str) {
     return str.substr(str.find(',') + 1, str.length());
 }
 
-inline string csvGetNext(string str) {
+inline string csvGetNext(const string& str) {
     return str.substr(0, str.find(','));
 }
 
-void UnpackRom(string romPath, string arm9) {
-    string command = string(NDS_TOOL) + " -x \"" + romPath + "\" -9 temp\\arm9.bin -7 temp\\arm7.bin -y9 temp\\y9.bin -y7 temp\\y7.bin -d data -y temp\\overlay -t temp\\banner.bin -h temp\\header.bin";
+void UnpackRom(const string& romPath, const string& arm9) {
+    string command = string(NDS_TOOL) + " -x \"" + romPath + R"(" -9 temp\arm9.bin -7 temp\arm7.bin -y9 temp\y9.bin -y7 temp\y7.bin -d data -y temp\overlay -t temp\banner.bin -h temp\header.bin)";
     cout << command << endl;
     system(command.c_str());
     
@@ -66,39 +66,42 @@ void UnpackRom(string romPath, string arm9) {
     changes.close();
     bin.close();
     
-    command = "copy files\\scr_seq.narc data\\a\\0\\1\\2";
+    command = R"(copy files\scr_seq.narc data\a\0\1\2)";
     system(command.c_str());
     
 }
 
 void UnpackFieldNarc() {
-    string command = string(KNARC) + " -d temp\\event_data -u data\\a\\0\\3\\2";
+    string command = string(KNARC) + R"( -d temp\event_data -u data\a\0\3\2)";
     system(command.c_str());
     
     //Updating some select files:
     for(const auto & file : directory_iterator(EVENT_DATA)) {
         string pstr = file.path().string();
         command = "copy " + file.path().string();
-        pstr = pstr.substr(pstr.find_last_of("\\") + 1, pstr.length());
+        pstr = pstr.substr(pstr.find_last_of('\\') + 1, pstr.length());
         command += " temp\\event_data\\" + pstr;
         system(command.c_str());
     }
 }
 
 void PackFieldNarc() {
-    string command = string(KNARC) + " -d temp\\event_data -p data\\a\\0\\3\\2";
+    string command = string(KNARC) + R"( -d temp\event_data -p data\a\0\3\2)";
     system(command.c_str());
 }
 
-void RepackRom(string output_name) {
+void RepackRom(const string& output_name) {
     cout << output_name << endl;
-    string command = string(NDS_TOOL) + " -c " + "\"" + string(OUT_PATH) + output_name + "\"" + " -9 temp\\arm9_.bin -7 temp\\arm7.bin -y9 temp\\y9.bin -y7 temp\\y7.bin -d data -y temp\\overlay -t temp\\banner.bin -h temp\\header.bin";
+    string command = string(NDS_TOOL) + " -c " + "\"" + string(OUT_PATH) + output_name + "\"" + R"( -9 temp\arm9_.bin -7 temp\arm7.bin -y9 temp\y9.bin -y7 temp\y7.bin -d data -y temp\overlay -t temp\banner.bin -h temp\header.bin)";
     system(command.c_str());
 }
 
 static void deleteDirectoryContents(const std::string& dir_path)
 {
-    for (const auto& entry : std::filesystem::directory_iterator(dir_path)) 
+    if(!std::filesystem::exists(dir_path)){
+        return;
+    }
+    for (const auto& entry : std::filesystem::directory_iterator(dir_path))
         std::filesystem::remove_all(entry.path());
 }
